+++ CHANGELOG.md
@@ -24,7 +24,7 @@
  - Other apps audio interruption on iOS
 - __Changes__
  - Deprecate event callbacks
- - Notification not visible anymore on lock screen
+ - Notification not visible by default on lock screen
  - Remove ticker property on Android
  - Remove unexpected back button handler
  - Remove support for wp8 platform